mv all out of the main
