+++ photon_stream/production/isdc/__init__.py
@@ -1,4 +1,4 @@
 from .qsub import qsub
 from . import dummy_qsub
 from . import qstat
-from .produce import produce
+from ._produce import produce
