+++ setup.py
@@ -40,6 +40,7 @@ setup(
     entry_points={'console_scripts': [
         'phs_extract_muons = photon_stream.muons.isdc_production.worker_node_main:main',
         'phs.isdc.obs.produce = photon_stream.production.isdc.produce:main',
+        'phs.isdc.obs.check = photon_stream.production.isdc.check:main',
     ]},
     zip_safe=False,
 )
